fixed EVERY BUG IN THE ACCOUNT PAGE !!! 1- dark mode bug 2- re-assigning the username after change it 3- converted the Change username page & Change password page from individual classes to a popup into the account page.
