+++ src/Main.java
@@ -4,8 +4,5 @@ import Utils.Generator;
 public class Main {
     public static void main(String[] args) {
         new HomePage();
-//        for(int i = 0; i < 100; i++){
-//            System.out.println(Generator.randomDateGen());
-//        }
     }
 }
